requirements.txt
Django==1.7.1
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
django-celery==3.1.16
django-extensions==1.4.8
djangorestframework==3.0.0
gevent==1.0.1
greenlet==0.4.5
grequests==0.2.0
kombu==3.0.24
pytz==2014.10
redis==2.10.3
requests==2.5.0
six==1.8.0
wsgiref==0.1.2
